# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Adunni Foods API
  description: |
    # Adunni Foods - Plantain Chips Ordering System
    
    A comprehensive REST API for managing plantain chips products, customer orders, and business operations.
    
    ## Features
    - **Product Management**: CRUD operations for plantain chips products
    - **Order Processing**: Customer order creation and management
    - **Admin Authentication**: Secure admin login system
    - **File Uploads**: Product image management via Cloudinary
    - **WhatsApp Integration**: Order notifications and communication
    - **Email Services**: Customer communication via SendGrid
    - **CSV Export**: Order data export functionality
    
    ## Authentication
    - JWT-based authentication for admin endpoints
    - Include `Authorization: Bearer <token>` header for protected routes
    
    ## Base URL
    - Development: `http://localhost:5000/api`
    - Production: `https://api.adunnifoods.com/api`
    
  version: 1.0.0
  contact:
    name: API Support
    email: adunnifoods.ltd@gmail.com
    url: https://adunnifoods.vercel.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.adunnifoods.com/api
    description: Production server
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: List All Products
      description: Retrieve a list of all available plantain chips products
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of products per page
        - in: query
          name: search
          schema:
            type: string
          description: Search products by name or description
      responses:
        '200':
          description: List of products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - Products
      summary: Create New Product
      description: Create a new plantain chips product (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
            example:
              name: "Spicy Plantain Chips"
              slug: "spicy-plantain-chips"
              description: "Delicious spicy plantain chips with a kick"
              price: 15.99
              stock: 100
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Product with this slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /orders:
    get:
      tags:
        - Orders
      summary: List All Orders
      description: Retrieve a list of all orders (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of orders per page
        - in: query
          name: status
          schema:
            type: string
            enum: [Pending, Packed, "Out for Delivery", Completed]
          description: Filter orders by status
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          description: Filter orders from this date (YYYY-MM-DD)
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          description: Filter orders until this date (YYYY-MM-DD)
      responses:
        '200':
          description: List of orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          description: Unauthorized - Invalid or missing token
    
    post:
      tags:
        - Orders
      summary: Create New Order
      description: Create a new customer order (Public endpoint)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              customerName: "John Doe"
              customerPhone: "2348144665646"
              address: "123 Main Street, Lagos, Nigeria"
              items:
                - product: "507f1f77bcf86cd799439011"
                  name: "Spicy Plantain Chips"
                  qty: 2
                  price: 15.99
              totalAmount: 31.98
              paymentMethod: "bank_transfer"
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    CreateProductRequest:
      type: object
      required:
        - name
        - slug
        - price
      properties:
        name:
          type: string
          minLength: 1
          description: Product name
          example: "Spicy Plantain Chips"
        slug:
          type: string
          minLength: 1
          description: URL-friendly product identifier
          example: "spicy-plantain-chips"
        description:
          type: string
          description: Product description
          example: "Delicious spicy plantain chips with a kick"
        price:
          type: number
          minimum: 0
          description: Product price in local currency
          example: 15.99
        stock:
          type: integer
          minimum: 0
          default: 0
          description: Available stock quantity
          example: 100
        images:
          type: array
          items:
            type: string
            format: uri
          description: Array of product image URLs
          example: ["https://res.cloudinary.com/djqfvpvsx/image/upload/v1/products/spicy-chips.jpg"]
...
    ProductResponse:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "Spicy Plantain Chips"
        slug:
          type: string
          example: "spicy-plantain-chips"
        description:
          type: string
          example: "Delicious spicy plantain chips with a kick"
        price:
          type: number
          example: 15.99
        stock:
          type: integer
          example: 100
        images:
          type: array
          items:
            type: string
            format: uri
          example: ["https://res.cloudinary.com/djqfvpvsx/image/upload/v1/products/spicy-chips.jpg"]
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
    ProductListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        pagination:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            total:
              type: integer
              example: 50
            pages:
              type: integer
              example: 3

    # Order Schemas
    OrderItem:
      type: object
      required:
        - product
        - name
        - qty
        - price
      properties:
        product:
          type: string
          description: Product ID reference
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          description: Product name
          example: "Spicy Plantain Chips"
        qty:
          type: integer
          minimum: 1
          description: Quantity ordered
          example: 2
        price:
          type: number
          minimum: 0
          description: Unit price
          example: 15.99
    CreateOrderRequest:
      type: object
      required:
        - customerName
        - customerPhone
        - address
        - items
        - totalAmount
        - paymentMethod
      properties:
        customerName:
          type: string
          minLength: 1
          description: Customer's full name
          example: "John Doe"
        customerPhone:
          type: string
          minLength: 6
          description: Customer's phone number
          example: "2348144665646"
        address:
          type: string
          minLength: 1
          description: Delivery address
          example: "123 Main Street, Lagos, Nigeria"
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'
          description: Array of ordered items
        totalAmount:
          type: number
          minimum: 0
          description: Total order amount
          example: 31.98
        paymentMethod:
          type: string
          enum: [bank_transfer, mobile_money, cash_on_delivery]
          description: Payment method
          example: "bank_transfer"
...
    OrderResponse:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        customerName:
          type: string
          example: "John Doe"
        customerPhone:
          type: string
          example: "2348144665646"
        address:
          type: string
          example: "123 Main Street, Lagos, Nigeria"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          example: 31.98
        paymentMethod:
          type: string
          enum: [bank_transfer, mobile_money, cash_on_delivery]
          example: "bank_transfer"
        status:
          type: string
          enum: [Pending, Packed, "Out for Delivery", Completed]
          example: "Pending"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
    CreateOrderResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Order created successfully"
        data:
          $ref: '#/components/schemas/OrderResponse'
        whatsappUrl:
          type: string
          format: uri
          description: WhatsApp URL for order confirmation
          example: "https://wa.me/2348144665646?text=Order%20confirmed%20for%20John%20Doe"
    OrderListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/OrderResponse'
        pagination:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            total:
              type: integer
              example: 25
            pages:
              type: integer
              example: 2

    # Upload Schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error message
          example: "Product not found"
        error:
          type: string
          description: Error type
          example: "NotFoundError"
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Validation failed"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field name with error
                example: "price"
              message:
                type: string
                description: Validation error message
                example: "Price must be a positive number"

# Security
security:
  - BearerAuth: []